+++ flowman-spec/src/main/scala/com/dimajix/flowman/spec/target/TargetSpec.scala
@@ -61,7 +61,8 @@ object TargetSpec extends TypeRegistry[TargetSpec] {
     new JsonSubTypes.Type(name = "sftpUpload", value = classOf[SftpUploadTargetSpec]),
     new JsonSubTypes.Type(name = "stream", value = classOf[StreamTargetSpec]),
     new JsonSubTypes.Type(name = "template", value = classOf[TemplateTargetSpec]),
-    new JsonSubTypes.Type(name = "validate", value = classOf[ValidateTargetSpec])
+    new JsonSubTypes.Type(name = "validate", value = classOf[ValidateTargetSpec]),
+    new JsonSubTypes.Type(name = "verify", value = classOf[VerifyTargetSpec])
 ))
 abstract class TargetSpec extends NamedSpec[Target] {
     @JsonProperty(value = "before", required=false) private var before:Seq[String] = Seq()