+++ CHANGELOG.md
@@ -28,7 +28,7 @@
 * Implement new testing capabilities  
 * Rename `update` mapping to `upsert` mapping, which better describes its functionality 
 * Introduce new `VALIDATE` phase, which is executed even before `CREATE` phase
-* Implement new `validate` target
+* Implement new `validate` and `verify` targets
 
 
 # Version 0.14.2 - 2020-10-12